+++ src/test/java/nextstep/ladder/domain/PlayerTest.java
@@ -1,8 +1,10 @@
-package nextstep.ladder.domain.common;
+package nextstep.ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
+import nextstep.ladder.domain.Player;
+import nextstep.ladder.domain.common.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
